+++ runelite-client/src/main/java/net/runelite/client/plugins/examine/ExaminePlugin.java
@@ -288,7 +288,7 @@ public class ExaminePlugin extends Plugin
 		// quantity is at least 1
 		quantity = Math.max(1, quantity);
 		int itemCompositionPrice = itemComposition.getPrice();
-		final int gePrice = itemManager.getItemPrice(id);
+		final long gePrice = itemManager.getItemPrice(id);
 		final long alchPrice = itemCompositionPrice <= 0 ? 0 : itemManager.getAlchValue(id);
 
 		if (gePrice > 0 || alchPrice > 0)
