+++ runelite-client/src/main/java/net/runelite/client/plugins/grounditems/GroundItem.java
@@ -40,7 +40,7 @@ class GroundItem
 	private WorldPoint location;
 	private int height;
 	private long haPrice;
-	private int gePrice;
+	private long gePrice;
 	private int offset;
 	private boolean tradeable;
 
@@ -49,7 +49,7 @@ class GroundItem
 		return haPrice * quantity;
 	}
 
-	int getGePrice()
+	long getGePrice()
 	{
 		return gePrice * quantity;
 	}
